+++ migrations/087_assessments__allow_issue_reporting__add.sql
@@ -1,0 +1,1 @@
+ALTER TABLE assessments ADD COLUMN allow_issue_reporting boolean DEFAULT FALSE;